workers/multimodal/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "datatree-multimodal"
version = "0.1.0"
description = "Local-only multimodal extraction sidecar for the datatree daemon (PDF text, OCR, ASR, notebooks)."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "Proprietary" }
authors = [
    { name = "datatree contributors" }
]
keywords = ["multimodal", "ocr", "pdf", "whisper", "local-first"]
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Operating System :: OS Independent",
    "Intended Audience :: Developers",
    "Topic :: Multimedia :: Sound/Audio :: Speech",
    "Topic :: Scientific/Engineering :: Image Recognition",
]

dependencies = [
    "pymupdf>=1.24.0",
    "pytesseract>=0.3.10",
    "Pillow>=10.0.0",
    "faster-whisper>=1.0.0",
    "numpy>=1.24.0",
    "pydantic>=2.5.0",
    "msgpack>=1.0.7",
]

[project.optional-dependencies]
docx = ["python-docx>=1.1.0"]
xlsx = ["openpyxl>=3.1.2"]
opencv = ["opencv-python-headless>=4.9.0"]
all = [
    "python-docx>=1.1.0",
    "openpyxl>=3.1.2",
    "opencv-python-headless>=4.9.0",
]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.4.0",
    "mypy>=1.8.0",
]

[project.scripts]
datatree-multimodal = "datatree_multimodal.__main__:main"

[project.urls]
Homepage = "https://example.invalid/datatree"
Repository = "https://example.invalid/datatree"

[tool.setuptools.packages.find]
where = ["."]
include = ["datatree_multimodal*"]
exclude = ["datatree_multimodal.tests*"]

[tool.setuptools.package-data]
datatree_multimodal = ["py.typed"]

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --strict-markers"
testpaths = ["datatree_multimodal/tests"]
asyncio_mode = "auto"
markers = [
    "slow: tests requiring large model downloads",
    "model_required: skip when model files are not present on disk",
]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "N", "SIM"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
strict = true
warn_unused_ignores = true
disallow_untyped_defs = true
ignore_missing_imports = true
